functions/clustering/requirements.txt
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
functions-framework>=3.0.0
flask>=2.3.0
